python_apps/image_search/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
pyspark>=3.4.0
pyarrow>=12.0.0
lancedb>=0.3.0
